+++ FuturaeDemo/AppDelegate.m
@@ -170,9 +170,68 @@ didReceiveNotificationResponse:(UNNotificationResponse *)response
 }
 
 #pragma mark - FTRNotificationDelegate
+- (void)unlockSDK:(void(^)(void))callback  {
+    UnlockMethodType method = FTRClient.sharedClient.getActiveUnlockMethods.firstObject.integerValue;
+    switch(method){
+        case UnlockMethodTypeBiometric:
+        case UnlockMethodTypeBiometricsOrPasscode:
+            [self unlockWithBiometrics:callback];
+            break;
+        case UnlockMethodTypeSDKPin:
+            [self unlockWithPIN:callback];
+            break;
+        default:
+            break;
+    }
+}
+
+- (void)unlockWithPIN:(void(^)(void))callback {
+    __weak __typeof(self) weakSelf = self;
+    PinViewController *vc = [[UIStoryboard storyboardWithName:@"Main" bundle:nil] instantiateViewControllerWithIdentifier:@"pinViewController"];
+    [vc setPinMode:@"input"];
+    [vc setDidFinishWithPinWithCallback:^(NSString * _Nullable pin) {
+        [FTRClient.sharedClient unlockWithSDKPin:pin callback:^(NSError * _Nullable error) {
+            [weakSelf.window.rootViewController dismissViewControllerAnimated:YES completion:nil];
+            callback();
+        }];
+    }];
+    [weakSelf.window.rootViewController presentViewController:vc animated:YES completion:nil];
+}
+
+- (void)unlockWithBiometrics:(void(^)(void))callback  {
+    [FTRClient.sharedClient unlockWithBiometrics:^(NSError * _Nullable error) {
+            callback();
+    } promptReason:@"UNLOCK WITH BIOMETRICS"];
+}
+
 ////////////////////////////////////////////////////////////////////////////////
 - (void)approveAuthenticationReceived:(nonnull NSDictionary *)authenticationInfo
 {
+    if(FTRClient.sharedClient.isLocked){
+        __weak __typeof(self) weakSelf = self;
+        UIAlertController *ac = [UIAlertController alertControllerWithTitle:@"SDK IS LOCKED"
+                                                                    message:@"You need to unlock to proceed"
+                                                             preferredStyle:UIAlertControllerStyleAlert];
+        
+        [ac addAction:[UIAlertAction actionWithTitle:@"UNLOCK"
+                                               style:UIAlertActionStyleDefault
+                                             handler:^(UIAlertAction * _Nonnull action) {
+            [weakSelf unlockSDK:^{
+                dispatch_async(dispatch_get_main_queue(), ^{
+                    [weakSelf approveAuthenticationReceived:authenticationInfo];
+                });
+            }];
+        }]];
+        [ac addAction:[UIAlertAction actionWithTitle:@"CANCEL"
+                                               style:UIAlertActionStyleDestructive
+                                             handler:^(UIAlertAction * _Nonnull action) {
+            //
+        }]];
+        
+        [self.window.rootViewController presentViewController:ac animated:NO completion:nil];
+        return;
+    }
+    
     NSLog(@"Received approve authentication: %@", authenticationInfo);
 
     // extra_info contains dynamic contextual information about this the authentication
@@ -189,25 +248,61 @@ didReceiveNotificationResponse:(UNNotificationResponse *)response
         }
     }
     NSNumber *sessionTimeout = authenticationInfo[@"session_timeout"];
+    NSArray<NSNumber *> *numbersChallenge = authenticationInfo[@"multi_numbered_challenge"];
     UIAlertController *ac = [UIAlertController alertControllerWithTitle:@"Approve"
                                                                 message:[NSString stringWithFormat: @"Would you like to approve the request? %@. \nSession timeout:\n %@ seconds.", extraInfoMsg, sessionTimeout.stringValue]
                                                          preferredStyle:UIAlertControllerStyleAlert];
+    
     [ac addAction:[UIAlertAction actionWithTitle:@"Approve" style:UIAlertActionStyleCancel handler:^(UIAlertAction * _Nonnull action) {
-        [[FTRClient sharedClient] approveAuthWithUserId:authenticationInfo[@"user_id"] sessionId:authenticationInfo[@"session_id"] extraInfo:authenticationInfo[@"extra_info"] callback:^(NSError * _Nullable error) {
-            if (error != nil) {
-                NSLog(@"Failed to approve: %@", error);
-                return;
-            }
-        }];
+        if(numbersChallenge){
+            [self pickMultiChallengeNumber:numbersChallenge callback:^(NSNumber *number) {
+                [[FTRClient sharedClient] approveAuthWithUserId:authenticationInfo[@"user_id"]
+                                                      sessionId:authenticationInfo[@"session_id"]
+                                                      extraInfo:authenticationInfo[@"extra_info"]
+                                              multiNumberChoice: number callback:^(NSError * _Nullable error) {
+                    if (error != nil) {
+                        NSLog(@"Failed to approve: %@", error);
+                        return;
+                    }
+                }];
+            }];
+        } else {
+            [[FTRClient sharedClient] approveAuthWithUserId:authenticationInfo[@"user_id"] sessionId:authenticationInfo[@"session_id"] extraInfo:authenticationInfo[@"extra_info"] callback:^(NSError * _Nullable error) {
+                if (error != nil) {
+                    NSLog(@"Failed to approve: %@", error);
+                    return;
+                }
+            }];
+        }
     }]];
     [ac addAction:[UIAlertAction actionWithTitle:@"Deny" style:UIAlertActionStyleDestructive handler:^(UIAlertAction * _Nonnull action) {
-        [[FTRClient sharedClient] rejectAuthWithUserId:authenticationInfo[@"user_id"] sessionId:authenticationInfo[@"session_id"] isFraud:NO extraInfo:authenticationInfo[@"extra_info"]  callback:^(NSError * _Nullable error) {
-            if (error != nil) {
-                NSLog(@"Failed to reject: %@", error);
-                return;
-            }
-        }];
+            [[FTRClient sharedClient] rejectAuthWithUserId:authenticationInfo[@"user_id"]
+                                                 sessionId:authenticationInfo[@"session_id"]
+                                                   isFraud:NO
+                                                 extraInfo:authenticationInfo[@"extra_info"]
+                                                  callback:^(NSError * _Nullable error) {
+                if (error != nil) {
+                    NSLog(@"Failed to reject: %@", error);
+                    return;
+                }
+            }];
     }]];
+    [self.window.rootViewController presentViewController:ac animated:NO completion:nil];
+}
+
+- (void)pickMultiChallengeNumber: (NSArray<NSNumber *> *)numbers callback:(void(^)(NSNumber *))callback {
+    UIAlertController *ac = [UIAlertController alertControllerWithTitle:@"Multi Number Challenge"
+                                                                message:@"Pick number"
+                                                         preferredStyle:UIAlertControllerStyleAlert];
+    
+    for (NSNumber *number in numbers) {
+        [ac addAction:[UIAlertAction actionWithTitle:[NSString stringWithFormat:@"%@", number]
+                                               style:UIAlertActionStyleDefault
+                                             handler:^(UIAlertAction * _Nonnull action) {
+            callback(number);
+        }]];
+    }
+    
     [self.window.rootViewController presentViewController:ac animated:NO completion:nil];
 }