+++ Package.swift
@@ -12,8 +12,8 @@ let package = Package(
     targets: [
         .binaryTarget(
           name: "FuturaeKit",
-          url: "https://github.com/Futurae-Technologies/ios-sdk/releases/download/v2.3.2/FuturaeKit-v2.3.2.xcframework.zip",
-          checksum: "61cd0d2f2fdab8c6da3d831f999154f97c77f1a891681061ba1ec1be5c7f9048"
+          url: "https://github.com/Futurae-Technologies/ios-sdk/releases/download/v2.3.3/FuturaeKit-v2.3.3.xcframework.zip",
+          checksum: "47c1c6372b1bb3e64b751175fafb105c3ea131c57f992243bc8aaaf878525b8f"
         )
     ]
 )